% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_seurat_to_viewable_sce.R
\name{convert_seurat_to_viewable_sce}
\alias{convert_seurat_to_viewable_sce}
\title{Convert Seurat to viewable SCE}
\usage{
convert_seurat_to_viewable_sce(
  so,
  assay = "Nanostring",
  fov_id = "fov_name",
  reductions = c()
)
}
\arguments{
\item{so}{Seurat object (already processed, annotated). Must be Seurat V5+!}

\item{assay}{Assay within Seurat object. For cosmx that's 'Nanostring'.
Expected to have counts and data layers.}

\item{fov_id}{Column in metadata with unique fov identifiers (Must be unique across all slides)}

\item{reductions}{Vector of reductions in Seurat object to copy across. If none specified, will take all.}
}
\value{
SingleCellExperiment object
}
\description{
Take a processed seurat object and convert to a SCE object with some features
for spatial browsing in iSEE. Mainly - two extra dimensional reductions.
}
\details{
\itemize{
\item local_spatial  : X,Y Coordinate of each cell within an FOV. Scaled to the 0-1 range for each fov.
\item global : Overall X,Y coorindates of each cell on the slide.
}

If not subsetted to fov / slide in iSEE, these will be overplotted. (No
concept of different slides)
}
\examples{

\dontrun{
sce <- convert_seurat_to_viewable_sce(so, fov_id = 'fov_name');
}
}
