% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_iSEE.R
\name{launch_iSEE_with_spatial_view}
\alias{launch_iSEE_with_spatial_view}
\title{Convert Seurat to viewable SCE}
\usage{
launch_iSEE_with_spatial_view(
  sce,
  main_cats = c(),
  plot_mode = "spatial",
  ecm = get_ecm(),
  max_cat = 40
)
}
\arguments{
\item{sce}{single cell experiment object to display (Suggest generating with convert_seurat_to_viewable_sce() )}

\item{main_cats}{Vector of metadata/cell information column names of
particular interest. The first one will end up plotted by default
(typically set to cluster), others brought up to the front of the drop down
selection for easy access.}

\item{plot_mode}{How to setup plot panels in interface. Currently only 'spatial' or NULL (iSEE defaults).}

\item{ecm}{Custom colourmap used in display (handy for sample colours. See iSEE docs.)}

\item{max_cat}{More than this many categoricals, and data will be treated as continuous. (default=40)}
}
\value{
None. Launches iSEE interface.
}
\description{
Take a processed seurat object and convert to a SCE object with some features
for spatial browsing in iSEE. Mainly - two extra dimensional reductions.
}
\details{
\itemize{
\item local_spatial  : X,Y Coordinate of each cell within an FOV. Scaled to the 0-1 range for each fov.
\item global : Overall X,Y coorindates of each cell on the slide.
}

If not subsetted to fov / slide in iSEE, these will be overplotted. (No
concept of different slides)
}
\examples{

\dontrun{
# Go
launch_iSEE_with_spatial_view(sce)

# make clusters the default grouping in various plots (and bring sample/fov up to top of lists)
launch_iSEE_with_spatial_view(sce, c("cluster","sample","fov"))
}
}
